requirements.txt
dnspython~=2.6.1
idna~=3.8
httpcore~=1.0.5
httpx~=0.27.2
anyio~=4.4.0
h11~=0.14.0
typing_extensions~=4.12.2
pillow~=10.2.0
pip~=24.2
filelock~=3.13.1
rich~=13.8.0
Pygments~=2.18.0
PyYAML~=6.0.2
sniffio~=1.3.1
click~=8.1.7
certifi~=2024.8.30
mdurl~=0.1.2
numpy~=1.26.3
setuptools~=65.5.0
sympy~=1.12
mpmath~=1.3.0
torch~=2.4.0+cu118
torchvision~=0.19.0+cu118
colorama~=0.4.6
fsspec~=2024.2.0
typer~=0.12.5
shellingham~=1.5.4
MarkupSafe~=2.1.5
fastapi~=0.112.2
starlette~=0.38.3
email_validator~=2.2.0
pydantic~=2.8.2
pydantic_core~=2.20.1
websockets~=13.0.1
uvicorn~=0.30.6
httptools~=0.6.1
watchfiles~=0.24.0
networkx~=3.2.1
python-multipart~=0.0.9
Jinja2~=3.1.3
torchaudio~=2.4.0+cu118
peewee~=3.17.6
python-dotenv~=1.0.1